Split ES controller into multiple sub controllers
